quaestio/requirements.txt
coverage==4.0.3
Django==1.9
djangorestframework==3.3.2
wheel==0.24.0
